requirements.txt
aiofiles==23.2.1
aiogram==3.10.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asyncpg==0.29.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.6.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cv==1.0.0
cycler==0.12.1
DAWG-Python==0.7.2
docopt==0.6.2
et-xmlfile==1.1.0
Flask==3.0.1
fonttools==4.43.1
frozenlist==1.4.1
googletrans==3.0.0
greenlet==3.0.3
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.7.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
magic-filter==1.0.12
MarkupSafe==2.1.4
matplotlib==3.8.0
multidict==6.0.5
numpy==1.26.1
opencv-python==4.8.1.78
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
psycopg==3.2.1
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pygame==2.5.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.1.1
PyQt6==6.6.1
PyQt6-Qt6==6.6.2
PyQt6-sip==13.6.0
pyTelegramBotAPI==4.20.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.32.3
rfc3986==1.5.0
setuptools==69.1.1
sip==6.8.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
telebot==0.0.5
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.2.2
Werkzeug==3.0.1
xlrd==2.0.1
XlsxWriter==3.1.9
yarl==1.9.4
